Assignment
Define the class DemoTest1 with the following methods:

A.	Static method addDemo ( ) that receives as arguments two objects of the class Demo and then builds and returns another object of the class Demo such that the value of each of its member variables is the sum of the values of the corresponding member variables of its arguments.

B.	Static method incrDemo2 that receives as argument an object of the class Demo and increments the value of each of its member variables by 5.

C.	method main that does the following:
1.	Declare object item of the class Demo.
2.	Read the values for the instance variables of object item, and set their values.
3.	Compute the average of the values of the instance variables of object item and print it.
4.	Declare object obj1 and set its instance variables to 5 and 7 respectively.
5.	Declare object obj2 and set its instance variables to 14 and 9 respectively.
6.	Create a third object named objR such that the value of each of its instance variables is the sum of the values of the corresponding instance variables of objects obj1 and obj2 by calling addDemo (  ).
7.	Increment the value of each instance variable of object obj1 by 5 by calling incrDemo(  ).
8.	Print the values of the instance variables of the objects objR and obj1.